+++ example/projection/spring-subscription-based-mongodb-projections/src/main/java/org/occurrent/example/eventstore/mongodb/spring/subscriptionprojections/SubscriptionProjectionsWithSpringAndMongoDBApplication.java
@@ -17,12 +17,10 @@
 package org.occurrent.example.eventstore.mongodb.spring.subscriptionprojections;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.cloudevents.CloudEvent;
 import org.occurrent.eventstore.api.blocking.EventStore;
 import org.occurrent.eventstore.mongodb.spring.blocking.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.spring.blocking.SpringBlockingMongoEventStore;
 import org.occurrent.mongodb.timerepresentation.TimeRepresentation;
-import org.occurrent.subscription.api.blocking.BlockingSubscription;
 import org.occurrent.subscription.api.blocking.BlockingSubscriptionPositionStorage;
 import org.occurrent.subscription.api.blocking.PositionAwareBlockingSubscription;
 import org.occurrent.subscription.mongodb.spring.blocking.SpringBlockingSubscriptionForMongoDB;
@@ -30,6 +28,7 @@ import org.occurrent.subscription.mongodb.spring.blocking.SpringBlockingSubscrip
 import org.occurrent.subscription.util.blocking.BlockingSubscriptionWithAutomaticPositionPersistence;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
 import org.springframework.data.mongodb.MongoTransactionManager;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -65,8 +64,9 @@ public class SubscriptionProjectionsWithSpringAndMongoDBApplication {
         return new SpringBlockingSubscriptionPositionStorageForMongoDB(mongoTemplate, "subscriptions");
     }
 
+    @Primary
     @Bean
-    public BlockingSubscription<CloudEvent> blockingSubscriptionWithAutomaticPositionPersistence(PositionAwareBlockingSubscription subscription, BlockingSubscriptionPositionStorage storage) {
+    public PositionAwareBlockingSubscription blockingSubscriptionWithAutomaticPositionPersistence(PositionAwareBlockingSubscription subscription, BlockingSubscriptionPositionStorage storage) {
         return new BlockingSubscriptionWithAutomaticPositionPersistence(subscription, storage);
     }