+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/BlockingSubscription.java
@@ -28,11 +28,8 @@ import java.util.function.Supplier;
  * and react to these events. Typically a subscription will forward the event to another piece of infrastructure such as
  * a message bus or to create views from the events (such as projections, sagas, snapshots etc).
  *
- * @param <T> The type of the {@link CloudEvent} that the subscription produce. It's common that subscriptions
- *            produce "wrappers" around {@code CloudEvent}'s that includes the subscription position if the event store
- *            doesn't maintain this.
  */
-public interface BlockingSubscription<T extends CloudEvent> {
+public interface BlockingSubscription {
 
     /**
      * Start listening to cloud events persisted to the event store using the supplied start position and <code>filter</code>.
@@ -45,7 +42,7 @@ public interface BlockingSubscription<T extends CloudEvent> {
      *                        were when the application was first started.
      * @param action          This action will be invoked for each cloud event that is stored in the EventStore.
      */
-    Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Supplier<StartAt> startAtSupplier, Consumer<T> action);
+    Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Supplier<StartAt> startAtSupplier, Consumer<CloudEvent> action);
 
 
     /**
@@ -56,7 +53,7 @@ public interface BlockingSubscription<T extends CloudEvent> {
      * @param startAt        The position to start the subscription from
      * @param action         This action will be invoked for each cloud event that is stored in the EventStore.
      */
-    default Subscription subscribe(String subscriptionId, SubscriptionFilter filter, StartAt startAt, Consumer<T> action) {
+    default Subscription subscribe(String subscriptionId, SubscriptionFilter filter, StartAt startAt, Consumer<CloudEvent> action) {
         return subscribe(subscriptionId, filter, () -> startAt, action);
     }
 
@@ -67,7 +64,7 @@ public interface BlockingSubscription<T extends CloudEvent> {
      * @param startAt        The position to start the subscription from
      * @param action         This action will be invoked for each cloud event that is stored in the EventStore.
      */
-    default Subscription subscribe(String subscriptionId, StartAt startAt, Consumer<T> action) {
+    default Subscription subscribe(String subscriptionId, StartAt startAt, Consumer<CloudEvent> action) {
         return subscribe(subscriptionId, null, startAt, action);
     }
 
@@ -78,7 +75,7 @@ public interface BlockingSubscription<T extends CloudEvent> {
      * @param filter         The filter to use to limit which events that are of interest from the EventStore.
      * @param action         This action will be invoked for each cloud event that is stored in the EventStore.
      */
-    default Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Consumer<T> action) {
+    default Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Consumer<CloudEvent> action) {
         return subscribe(subscriptionId, filter, StartAt.now(), action);
     }
 
@@ -88,7 +85,7 @@ public interface BlockingSubscription<T extends CloudEvent> {
      * @param subscriptionId The id of the subscription, must be unique!
      * @param action         This action will be invoked for each cloud event that is stored in the EventStore.
      */
-    default Subscription subscribe(String subscriptionId, Consumer<T> action) {
+    default Subscription subscribe(String subscriptionId, Consumer<CloudEvent> action) {
         return subscribe(subscriptionId, null, StartAt.now(), action);
     }