+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/PositionAwareBlockingSubscription.java
@@ -17,15 +17,15 @@
 package org.occurrent.subscription.api.blocking;
 
 import io.cloudevents.CloudEvent;
-import org.occurrent.subscription.CloudEventWithSubscriptionPosition;
+import org.occurrent.subscription.PositionAwareCloudEvent;
 import org.occurrent.subscription.SubscriptionPosition;
 
 /**
- * A {@link BlockingSubscription} that produces {@link CloudEventWithSubscriptionPosition} compatible {@link CloudEvent}'s.
+ * A {@link BlockingSubscription} that produces {@link PositionAwareCloudEvent} compatible {@link CloudEvent}'s.
  * This is useful for subscribers that want to persist the subscription position for a given subscription if the event store doesn't
  * maintain the position for subscriptions.
  */
-public interface PositionAwareBlockingSubscription extends BlockingSubscription<CloudEventWithSubscriptionPosition> {
+public interface PositionAwareBlockingSubscription extends BlockingSubscription {
 
     /**
      * The global subscription position might be e.g. the wall clock time of the server, vector clock, number of events consumed etc.