+++ subscription/api/reactor/src/main/java/org/occurrent/subscription/api/reactor/ReactorSubscription.java
@@ -26,28 +26,24 @@ import reactor.core.publisher.Flux;
  * Common interface for reactor (reactive) subscriptions. The purpose of a subscription is to read events from an event store
  * and react to these events. Typically a subscription will forward the event to another piece of infrastructure such as
  * a message bus or to create views from the events (such as projections, sagas, snapshots etc).
- *
- * @param <T> The type of the {@link CloudEvent} that the subscription produce. It's common that subscriptions
- *            produce "wrappers" around {@code CloudEvent}'s that includes the subscription position if the event store
- *            doesn't maintain this.
  */
-public interface ReactorSubscription<T extends CloudEvent> {
+public interface ReactorSubscription {
 
     /**
      * Stream events from the event store as they arrive and provide a function which allows to configure the
-     * {@link T} that is used. Use this method if want to start streaming from a specific
+     * {@link SubscriptionFilter} that is used. Use this method if want to start streaming from a specific
      * position.
      *
-     * @return A {@link Flux} with cloud events which also includes the {@link SubscriptionPosition} that can be used to resume the stream from the current position.
+     * @return A {@link Flux} with cloud events.
      */
-    Flux<T> subscribe(SubscriptionFilter filter, StartAt startAt);
+    Flux<CloudEvent> subscribe(SubscriptionFilter filter, StartAt startAt);
 
     /**
      * Stream events from the event store as they arrive but filter only events that matches the <code>filter</code>.
      *
      * @return A {@link Flux} with cloud events which also includes the {@link SubscriptionPosition} that can be used to resume the stream from the current position.
      */
-    default Flux<T> subscribe(SubscriptionFilter filter) {
+    default Flux<CloudEvent> subscribe(SubscriptionFilter filter) {
         return subscribe(filter, StartAt.now());
     }
 
@@ -57,7 +53,7 @@ public interface ReactorSubscription<T extends CloudEvent> {
      *
      * @return A {@link Flux} with cloud events which also includes the {@link SubscriptionPosition} that can be used to resume the stream from the current position.
      */
-    default Flux<T> subscribe(StartAt startAt) {
+    default Flux<CloudEvent> subscribe(StartAt startAt) {
         return subscribe(null, startAt);
     }
 
@@ -66,7 +62,7 @@ public interface ReactorSubscription<T extends CloudEvent> {
      *
      * @return A {@link Flux} with cloud events which also includes the {@link SubscriptionPosition} that can be used to resume the stream from the current position.
      */
-    default Flux<T> subscribe() {
+    default Flux<CloudEvent> subscribe() {
         return subscribe(null, StartAt.now());
     }
 }